Tests/LibWeb/Text/expected/canvas/webgl-present-with-bound-offscreen-framebuffer.txt

FIXME: This test relies on having a GPU backend enabled, which it is not in run-tests mode :(
